Rewrites the pending plugin to simplify things greatly.
The plugin now just takes a trace level from clients and uses that plus gas limit for a transaction to decide whether to trace a pending transaction or not.  If any connected client decides to trace a pending transaction then all transactions with a gas price greater than 21,000 will get traced and the results of those traces will be sent to connected clients.

Also consolidates the API so the same shape data is always sent, just with or without trace details.